Switch to using tianocore/edk2 instead of ionescu007/edk2
VisualUefi refers changeset f4c5200926df8aa8e4ede24fcca288be2150acb4 from edk2. This changeset was not (at the time of writing) contained in a branch in ionescu007's fork but it can be found in the original tianocore's repo's master.

Since git can fail to clone/checkout revisions not currently contained in a branch and we do not need custom code from the fork, it was decided it would be best to just use the original repository.